requirements.txt
amqp==1.4.6
anyjson==0.3.3
audioread==1.2.1
billiard==3.3.0.19
boto==2.36.0
bugsnag==2.3.0
celery==3.1.17
decorator==3.4.0
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.7.5
-e git://github.com/yotamofek/google-api-python-client.git@python3#egg=google_api_python_client-origin_python3
greenlet==0.4.5
httplib2==0.9
kombu==3.0.24
oauth2client==1.4.2
Pillow==2.7.0
psycopg2==2.6
pyasn1==0.1.7
pyasn1-modules==0.0.5
pytz==2014.10
redis==2.10.3
requests==2.5.3
rsa==3.1.4
simplejson==3.6.5
six==1.9.0
SQLAlchemy==0.9.8
static3==0.5.1
uritemplate==0.6
uWSGI==2.0.9
WebOb==1.4
Werkzeug==0.10.1
ws4py==0.3.4
